requirements/base.txt
openai>=1.30.0
gradio>=4.28.0
ffmpeg-python>=0.2.0
Pillow>=10.0.0
numpy>=1.26.0
pydub>=0.25.1
google-cloud-storage>=2.14.0
python-dotenv>=1.0.1
tenacity>=8.2.3
